Comunication reworked. MainScreen reworked.

+++ Statemachine_ST7783_ADC/App/CSharp/LoggerVS08/Logger/Logger/LoggerTabPage.cs
@@ -240,7 +240,7 @@ namespace Logger {
 			}
 		}
 
-		public int SensorId { get; set; }
+		public byte SensorId { get; set; }
 	}
 
 	public class ProgressBar : ToolStripProgressBar {
